Use set -e to fail if rsync doesn't complete
If rsync can't connect over SSH, it exits with a zero
exit code, which denotes success, and CodeShip thinks
the deployment completed successfully. I suspect that
since rsync is executing the SSH command, when SSH fails
the error is with SSH inside an rsync wrapper. rsync
does not pass the SSH error code through and return it.
rsync just continues on its happy way, and we get a zero
exit code (i.e. success) from rsync.

I've tried unsuccessfully to detect the SSH-failure exit code,
and to grep an error message from rsync.

Using set -e does the trick, though. Not sure how it's
detecting an error I couldn't pick up, but I guess
that's what it's for.

I've also tried using set -e for the entire script, but
this led to early failures without any debugging info,
perhaps because there are no-zero exit codes inherent to
the git-related checks we do in this script.